fix: :recycle: improve header login
