+++ easystore-app/app/[locale]/components/organisms/login/HeaderLogin.tsx
@@ -1,9 +1,30 @@
-import { LanguageButton } from '@components/atoms/shared/ButtonLanguage';
+'use client';
 
-export default function HeaderLogin() {
+import React from 'react';
+import LogoLogin from '@components/atoms/login/LogoLogin';
+import { LanguageButton } from '@components/atoms/shared/ButtonLanguage';
+import { useTranslations } from 'next-intl';
+
+export default function Header() {
+  const t = useTranslations('Login');
+
   return (
-    <header className="bg-background sticky top-0 z-50 flex h-12 items-center justify-end px-3 sm:h-14 sm:px-6 md:h-16 md:px-10 lg:h-20 lg:px-16">
-      <LanguageButton />
+    <header className="bg-background relative mt-4 px-4 py-6 sm:px-8 md:px-16 lg:px-32">
+      <div className="absolute top-4 right-4 sm:top-6 sm:right-8 md:top-6 md:right-16">
+        <LanguageButton />
+      </div>
+
+      <div className="flex flex-col items-center justify-center space-y-0 text-center sm:flex-row sm:space-y-0 sm:space-x-4 sm:text-left">
+        <LogoLogin />
+        <div>
+          <h1 className="text-title text-2xl font-extrabold sm:text-3xl md:text-4xl">
+            {t('welcomeBack')}
+          </h1>
+          <p className="text-text text-primary text-sm sm:text-base md:text-lg">
+            {t('loginMessage')}
+          </p>
+        </div>
+      </div>
     </header>
   );
 }
