+++ easystore-app/app/[locale]/components/organisms/login/MainLogin.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import { LoginForm } from '@components/molecules/login/LoginForm';
-import HeaderForm from '@components/molecules/login/HeaderForm';
 
 export default function MainLogin() {
   const handleLogin = (data: { email: string; password: string }) => {
@@ -10,7 +9,6 @@ export default function MainLogin() {
 
   return (
     <main className="mb-10 flex flex-col items-center justify-center px-4 sm:px-8 md:px-16 lg:px-32">
-      <HeaderForm />
       <LoginForm onSubmit={handleLogin} />
     </main>
   );
